refactor(e2e): merge 5 real-cli probes into harness-real-cli.mjs
5 standalone real-cli probes (cwd, new-session, reopen, import,
switch) consolidated into a single harness with --only/--skip
case selection. 4 cases share a single Electron launch
(saves ~45s over standalone runs); 1 case keeps own launches
(reopen-resume needs cross-restart with shared userDataDir).

Deleted: probe-real-{cwd-projects-claude,new-session-chat,
reopen-resume,import-resume,switch-session-keeps-chat}.mjs,
plus obsolete dogfood-probe-happy-path.mjs.

Harness wall time: 132s (was ~177s standalone, build excluded).
Net LOC: -2441 / +790 = -1651.

Per feedback_e2e_prefer_harness.md + feedback_local_e2e_only.md.
